First revision of optimization algorithms
While going through this I had the idea that the a reason why it may be hard to achive solutions when starting with all BEBs in their wait queues is because that scenario will always fail. Because of this, there should be a solution that is changed and maintained. Only a snapshot of the solution should be made when it is in the solution space.