Droge edits 01/14/24 (#40)
* Starting to reserve capital letters for sets

* working on problem desription

* Update problem description

* Working on optimization problem

This commit contains edits up to the objective function section

* Working on objective function

This commit has edits up to the constraints section

* Complete optimization problem section

* working on SA section introduction

* bug fixes

* Updating generators

* Update purge

* Update slide visit

* Update new charger

* Update new-window and wait generator functions

* edits up to perturb schedule

* working on optimization algorithms

* underline section to be removed

* Make SA algorithms match code more closely

* First revision of optimization algorithms

While going through this I had the idea that the a reason why it may be hard to achive solutions when starting with all BEBs in their wait queues is because that scenario will always fail. Because of this, there should be a solution that is changed and maintained. Only a snapshot of the solution should be made when it is in the solution space.

* Added line to return original visit if generator fails

* Added footnote

* Second go over introduction

* Second pass at optimization problem section

* Up to page 15 on second round of edits

* fix bugs

* update citiations

* finish second edits of simulated annealing section

* finish second edits optimization problem

* update notation for placing original queue state

* update notation for placing original queue state